[iOS] Volume Detect Code Appdelegate -> plugin.swift
