+++ example/ios/Runner/AppDelegate.swift
@@ -1,71 +1,13 @@
 import UIKit
 import Flutter
-import AVFoundation
-
-enum ChannelName {
-    static let volume = "flutter.moum.hardware_buttons.volume"
-}
 
 @UIApplicationMain
-@objc class AppDelegate: FlutterAppDelegate, FlutterStreamHandler {
-    
-    private var volumeLevel: Float = 0.0
-    var eventSink: FlutterEventSink?
-    
+@objc class AppDelegate: FlutterAppDelegate {
     override func application(_ application: UIApplication,
                               didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
         
         GeneratedPluginRegistrant.register(with: self)
-        
-        guard let controller = window?.rootViewController as? FlutterViewController else {
-            fatalError("rootViewController is not type FlutterViewController")
-        }
-        
-        let volumeChannel = FlutterEventChannel(name: ChannelName.volume,
-                                                binaryMessenger: controller.binaryMessenger)
-        volumeChannel.setStreamHandler(self)
-        
         return super.application(application,
                                  didFinishLaunchingWithOptions: launchOptions)
-    }
-    
-    func onListen(withArguments arguments: Any?,
-                  eventSink events: @escaping FlutterEventSink) -> FlutterError? {
-        self.eventSink = events
-        let audioSession = AVAudioSession.sharedInstance()
-        do {
-            try audioSession.setActive(true)
-            audioSession.addObserver(self,
-                                     forKeyPath: "outputVolume",
-                                     options: .new,
-                                     context: nil)
-            volumeLevel = audioSession.outputVolume
-        } catch {
-            print("error")
-        }
-        return nil
-    }
-    
-    override public func observeValue(forKeyPath keyPath: String?,
-                                      of object: Any?,
-                                      change: [NSKeyValueChangeKey: Any]?,
-                                      context: UnsafeMutableRawPointer?) {
-        if keyPath == "outputVolume" {
-            let audioSession = AVAudioSession.sharedInstance()
-            if audioSession.outputVolume > volumeLevel {
-                eventSink?(24)
-            }
-            if audioSession.outputVolume < volumeLevel {
-                eventSink?(25)
-            }
-            volumeLevel = audioSession.outputVolume
-        }
-    }
-    
-    func onCancel(withArguments arguments: Any?) -> FlutterError? {
-        NotificationCenter.default.removeObserver(self,
-                                                  forKeyPath: "outputVolume")
-        eventSink = nil
-        return nil
     }
 }
